Replacing Encrpyt.cpp with Encrpyt.c, Adding permissions check and moving asking for text encryption to functions.
Replacing Encrpyt.cpp with Encrpyt.c so that gcc can compile it. This changes *nothing* in the file, nor does it cause problems. If I am wrong, Revert this change.

Moving asking if the user wants to first encrypt their text message they wish to hide into a function, so as to save repeating a large chunk of code.

Add permissions checks for if the program can write (and read/execute) to the folder, since I believe I had problems with one program failing to write if I was using the shared folder between a VM and my main OS, where I was editing and had the directory open.
If this never matters, that's OK too. Just good to check.